+++ src/services/axios/sensitivityCalculator/sensCalHelpers.ts
@@ -27,7 +27,7 @@ const sensCalHelpers = {
      * @param precision the number of d.p. to display the result to
      * @returns {object} the sensitivity as an object with the correct units and precision // the sensitivity as a string with the correct units and precision
      * **/
-    convertSensitivityToDisplayValue(sensitivity: number, precision = 2): ValueUnitPair {
+    convertReturnedSensitivityToDisplayValue(sensitivity: number, precision = 2): ValueUnitPair {
       // TODO: add tests (cypress?)
       if (typeof sensitivity === 'number') {
         if (sensitivity < 1e3) {
@@ -97,7 +97,7 @@ const sensCalHelpers = {
     },
     /**
      * Converts a minor and major beam in degrees (as returned by the backend)
-     * into a formatted string in arcsecs eg '4.6" x 7.9"'
+     * into a formatted string in arcsec eg '4.6" x 7.9"'
      *
      * @param beam_min_scaled in degrees
      * @param beam_maj_scaled in degrees
@@ -150,9 +150,17 @@ const sensCalHelpers = {
       return bandwidthValue * unitMap[bandwidthUnits];
     },
     convertBandwidthToHz(bandwidthValue, bandwidthUnits): number {
-      const unitMap = [0, 1000000000, 1000000, 1000, 1]; // Spacer, GHz, MHz, KHz, Hz
-      if (bandwidthUnits < 1 || bandwidthUnits > unitMap.length) {
-        throw new Error(`Invalid bandwidth unit: ${bandwidthUnits}`);
+      if (typeof bandwidthUnits === 'number') {
+        bandwidthUnits = OBSERVATION.Units.find(item => item.value === bandwidthUnits)?.label;
+      }
+      const unitMap: { [key: string]: number } = {
+        GHz: 1000000000,
+        MHz: 1000000,
+        KHz: 1000,
+        Hz: 1
+      };
+      if (!unitMap[bandwidthUnits]) {
+        throw new Error('Invalid bandwidth unit');
       }
       return bandwidthValue * unitMap[bandwidthUnits];
     }